move wav2 to server.py
